pyproject.toml
[tool.poetry]
name = "social-media-app-backend"
version = "0.1.0"
description = ""
authors = ["TianxiangR <ren644120622@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
pillow = "^10.2.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
